+++ cmd/tempo/app/overrides_validation.go
@@ -2,12 +2,14 @@ package app
 
 import (
 	"fmt"
+	"time"
 
 	"golang.org/x/exp/slices"
 
 	"github.com/grafana/tempo/modules/generator"
 	"github.com/grafana/tempo/modules/overrides/userconfigurable/api"
 	"github.com/grafana/tempo/modules/overrides/userconfigurable/client"
+	filterconfig "github.com/grafana/tempo/pkg/spanfilter/config"
 )
 
 type overridesValidator struct {
@@ -45,6 +47,20 @@ func (v *overridesValidator) Validate(limits *client.Limits) error {
 			if !slices.Contains(generator.SupportedProcessors, p) {
 				return fmt.Errorf("metrics_generator.processor \"%s\" is not a known processor, valid values: %v", p, generator.SupportedProcessors)
 			}
+		}
+	}
+
+	if filterPolicies, ok := limits.GetMetricsGenerator().GetProcessor().GetSpanMetrics().GetFilterPolicies(); ok {
+		for _, fp := range filterPolicies {
+			if err := filterconfig.ValidateFilterPolicy(fp); err != nil {
+				return err
+			}
+		}
+	}
+
+	if collectionInterval, ok := limits.GetMetricsGenerator().GetCollectionInterval(); ok {
+		if collectionInterval < 15*time.Second || collectionInterval > 5*time.Minute {
+			return fmt.Errorf("metrics_generator.collection_interval \"%s\" is outside acceptable range of 15s to 5m", collectionInterval.String())
 		}
 	}