+++ cmd/tempo/app/overrides_validation_test.go
@@ -3,6 +3,7 @@ package app
 import (
 	"fmt"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 
@@ -10,6 +11,7 @@ import (
 	"github.com/grafana/tempo/modules/distributor/forwarder"
 	"github.com/grafana/tempo/modules/generator"
 	"github.com/grafana/tempo/modules/overrides/userconfigurable/client"
+	filterconfig "github.com/grafana/tempo/pkg/spanfilter/config"
 )
 
 func Test_overridesValidator(t *testing.T) {
@@ -61,6 +63,68 @@ func Test_overridesValidator(t *testing.T) {
 				},
 			},
 			expErr: fmt.Sprintf("metrics_generator.processor \"span-span\" is not a known processor, valid values: %v", generator.SupportedProcessors),
+		},
+		{
+			name: "filter policies",
+			cfg:  Config{},
+			limits: client.Limits{
+				Forwarders: &[]string{},
+				MetricsGenerator: &client.LimitsMetricsGenerator{Processor: &client.LimitsMetricsGeneratorProcessor{SpanMetrics: &client.LimitsMetricsGeneratorProcessorSpanMetrics{FilterPolicies: &[]filterconfig.FilterPolicy{{
+					Include: &filterconfig.PolicyMatch{
+						MatchType: filterconfig.Strict,
+						Attributes: []filterconfig.MatchPolicyAttribute{
+							{
+								Key:   "span.kind",
+								Value: "SPAN_KIND_SERVER",
+							},
+						},
+					},
+				}}}}},
+			},
+		},
+		{
+			name: "filter policies - invalid",
+			cfg:  Config{},
+			limits: client.Limits{
+				Forwarders: &[]string{},
+				MetricsGenerator: &client.LimitsMetricsGenerator{Processor: &client.LimitsMetricsGeneratorProcessor{SpanMetrics: &client.LimitsMetricsGeneratorProcessorSpanMetrics{FilterPolicies: &[]filterconfig.FilterPolicy{
+					{
+						Include: &filterconfig.PolicyMatch{
+							MatchType: "invalid",
+						},
+					},
+				}}}},
+			},
+			expErr: "invalid include policy: invalid match type: invalid",
+		},
+		{
+			name: "metrics_generator.collection_interval valid",
+			cfg:  Config{},
+			limits: client.Limits{
+				MetricsGenerator: &client.LimitsMetricsGenerator{
+					CollectionInterval: &client.Duration{Duration: 60 * time.Second},
+				},
+			},
+		},
+		{
+			name: "metrics_generator.collection_interval minimum",
+			cfg:  Config{},
+			limits: client.Limits{
+				MetricsGenerator: &client.LimitsMetricsGenerator{
+					CollectionInterval: &client.Duration{Duration: 1 * time.Second},
+				},
+			},
+			expErr: "metrics_generator.collection_interval \"1s\" is outside acceptable range of 15s to 5m",
+		},
+		{
+			name: "metrics_generator.collection_interval maximum",
+			cfg:  Config{},
+			limits: client.Limits{
+				MetricsGenerator: &client.LimitsMetricsGenerator{
+					CollectionInterval: &client.Duration{Duration: 10 * time.Minute},
+				},
+			},
+			expErr: "metrics_generator.collection_interval \"10m0s\" is outside acceptable range of 15s to 5m",
 		},
 	}